+++ src/services/lease-service/tests/routes/contacts.test.ts
@@ -8,7 +8,7 @@ import { routes } from '../../routes/contacts'
 import * as tenantLeaseAdapter from '../../adapters/xpand/tenant-lease-adapter'
 import * as xPandSoapAdapter from '../../adapters/xpand/xpand-soap-adapter'
 import * as applicationProfileAdapter from '../../adapters/application-profile-adapter'
-import * as applicationProfileService from '../../update-or-create-application-profile'
+import * as applicationProfileService from '../../create-or-update-application-profile'
 import * as factories from '../../tests/factories'
 
 const app = new Koa()
